Per-db option routes.
The options per-db are only the sorting, and they were already fixed.
What this is for is so that I can do a reverse-proxy on the non-private
dbs ONLY from www.katspace.org and they can serve as bookmark pages for
the public, while I can still have my private bookmarks on localhost.
But in order to do that, people need to be able to change their opts
while under the /db/foo hierarchy, rather than using /opt.

Hence the new per-db option routes.
The session cookie which stores the information is still a top-level one.