Added ability to add new bookmark files.
New form to get "bookmark" info, which will then save a YAML file in a
designated directory. This doesn't add anything to the database, and it
doesn't create a thumbnail file. It probably should, but I thought it
best to get the first working draft checked in.

This may need to be renamed, but I'm not sure as to what.
It certainly needs to be documented, which it isn't.