Redirection and logging.
Redirection appears not to play nice with a reverse proxy.
There might be some way of getting it to work properly, but it's
easier to just not use it, considering the amount of time I spent trying
to debug it (which is why I added the logging).
Now, when one changes /opt settings, instead of redirecting back to the
page you came from, you get shown a page with the new settings, and a
link to go back to where you came from. That's not user-unfriendly,
because at least it's giving you feedback that clicking the "Set" button
actually did something.