Figured out a way of using multiple config files.
It's a bit of a hack, but it works.
Used the %ENV hack described in
http://stackoverflow.com/questions/17594352/how-to-pass-arguments-from-an-plack-app-to-an-mojolicious-app-enabled-in-builder
plus I needed to have a separate script defined in the script dir
which set this different ENV variable, so, yes, it is a bit of a hack.
The way that Mojolicious works prevented something more elegant.